Refactor main.rs into modules for maintainability
Split the 965-line main.rs into focused modules:
- api.rs: OpenAI and Anthropic API structs + is_anthropic()
- config.rs: Config struct, load_config(), config_path(), resolve()
- command.rs: extract_command(), detect_destructive()
- explain.rs: render_explanation()
- wizard.rs: Provider, PROVIDERS, run_config_wizard()

All 24 tests pass with tests colocated in their respective modules.
