+++ src/wizard.rs
@@ -1,0 +1,218 @@
+use colored::Colorize;
+use dialoguer::{Confirm, Input, Password, Select};
+
+use crate::config::config_path;
+
+pub struct Provider {
+    pub name: &'static str,
+    pub api_base: &'static str,
+    pub needs_api_key: bool,
+    pub suggested_models: &'static [&'static str],
+}
+
+pub const PROVIDERS: &[Provider] = &[
+    Provider {
+        name: "OpenAI",
+        api_base: "https://api.openai.com/v1",
+        needs_api_key: true,
+        suggested_models: &["gpt-4o-mini", "gpt-4o", "gpt-4.1-mini", "gpt-4.1-nano"],
+    },
+    Provider {
+        name: "Anthropic",
+        api_base: "https://api.anthropic.com/v1",
+        needs_api_key: true,
+        suggested_models: &[
+            "claude-sonnet-4-6",
+            "claude-haiku-4-5-20251001",
+            "claude-opus-4-6",
+        ],
+    },
+    Provider {
+        name: "Ollama",
+        api_base: "http://localhost:11434/v1",
+        needs_api_key: false,
+        suggested_models: &["qwen3.5", "gemma3:4b", "llama3.2", "mistral"],
+    },
+    Provider {
+        name: "Groq",
+        api_base: "https://api.groq.com/openai/v1",
+        needs_api_key: true,
+        suggested_models: &[
+            "llama-3.3-70b-versatile",
+            "gemma2-9b-it",
+            "mixtral-8x7b-32768",
+        ],
+    },
+    Provider {
+        name: "Together AI",
+        api_base: "https://api.together.xyz/v1",
+        needs_api_key: true,
+        suggested_models: &[
+            "meta-llama/Llama-3-70b-chat-hf",
+            "mistralai/Mixtral-8x7B-Instruct-v0.1",
+        ],
+    },
+    Provider {
+        name: "OpenRouter",
+        api_base: "https://openrouter.ai/api/v1",
+        needs_api_key: true,
+        suggested_models: &[
+            "openai/gpt-4o-mini",
+            "anthropic/claude-sonnet-4-6",
+            "meta-llama/llama-3-70b-instruct",
+        ],
+    },
+    Provider {
+        name: "LM Studio",
+        api_base: "http://localhost:1234/v1",
+        needs_api_key: false,
+        suggested_models: &["qwen3.5", "gemma3:4b"],
+    },
+    Provider {
+        name: "vLLM",
+        api_base: "http://localhost:8000/v1",
+        needs_api_key: false,
+        suggested_models: &["meta-llama/Llama-3-8b-chat-hf"],
+    },
+    Provider {
+        name: "LocalAI",
+        api_base: "http://localhost:8080/v1",
+        needs_api_key: false,
+        suggested_models: &["gpt-4o-mini"],
+    },
+];
+
+pub fn run_config_wizard() {
+    eprintln!("{}", "yaak configuration wizard".bold());
+    eprintln!("{}", "─".repeat(40).dimmed());
+    eprintln!();
+
+    // 1. Select provider
+    let provider_names: Vec<&str> = PROVIDERS.iter().map(|p| p.name).collect();
+    let provider_idx = Select::new()
+        .with_prompt("Select your API provider")
+        .items(&provider_names)
+        .default(0)
+        .interact()
+        .unwrap_or_else(|_| {
+            eprintln!("{}", "Aborted.".dimmed());
+            std::process::exit(0);
+        });
+
+    let provider = &PROVIDERS[provider_idx];
+    eprintln!();
+
+    // 2. Select model
+    let mut model_options: Vec<String> = provider
+        .suggested_models
+        .iter()
+        .map(|m| m.to_string())
+        .collect();
+    model_options.push("Enter custom model name".to_string());
+
+    let model_idx = Select::new()
+        .with_prompt("Select a model")
+        .items(&model_options)
+        .default(0)
+        .interact()
+        .unwrap_or_else(|_| {
+            eprintln!("{}", "Aborted.".dimmed());
+            std::process::exit(0);
+        });
+
+    let model = if model_idx == model_options.len() - 1 {
+        Input::<String>::new()
+            .with_prompt("Enter model name")
+            .interact_text()
+            .unwrap_or_else(|_| {
+                eprintln!("{}", "Aborted.".dimmed());
+                std::process::exit(0);
+            })
+    } else {
+        model_options[model_idx].clone()
+    };
+    eprintln!();
+
+    // 3. API key (if needed)
+    let api_key = if provider.needs_api_key {
+        let key = Password::new()
+            .with_prompt("Enter your API key")
+            .interact()
+            .unwrap_or_else(|_| {
+                eprintln!("{}", "Aborted.".dimmed());
+                std::process::exit(0);
+            });
+        if key.is_empty() {
+            eprintln!(
+                "{} API key is required for {}",
+                "warning:".yellow().bold(),
+                provider.name
+            );
+            std::process::exit(1);
+        }
+        eprintln!();
+        Some(key)
+    } else {
+        None
+    };
+
+    // Build config TOML
+    let mut config_content = String::new();
+    config_content.push_str("# yaak configuration — generated by `yaak --config`\n\n");
+    config_content.push_str(&format!("api_base = \"{}\"\n", provider.api_base));
+    if let Some(key) = &api_key {
+        config_content.push_str(&format!("api_key = \"{}\"\n", key));
+    }
+    config_content.push_str(&format!("model = \"{}\"\n", model));
+
+    // Write config file
+    let path = config_path();
+    if let Some(parent) = path.parent() {
+        if let Err(e) = std::fs::create_dir_all(parent) {
+            eprintln!(
+                "{} Failed to create config directory: {}",
+                "error:".red().bold(),
+                e
+            );
+            std::process::exit(1);
+        }
+    }
+
+    // Warn if overwriting
+    if path.exists() {
+        let overwrite = Confirm::new()
+            .with_prompt(format!(
+                "Config already exists at {}. Overwrite?",
+                path.display()
+            ))
+            .default(false)
+            .interact()
+            .unwrap_or(false);
+        if !overwrite {
+            eprintln!("{}", "Aborted.".dimmed());
+            std::process::exit(0);
+        }
+    }
+
+    if let Err(e) = std::fs::write(&path, &config_content) {
+        eprintln!("{} Failed to write config: {}", "error:".red().bold(), e);
+        std::process::exit(1);
+    }
+
+    eprintln!("{}", "─".repeat(40).dimmed());
+    eprintln!(
+        "{} Config written to {}",
+        "✓".green().bold(),
+        path.display().to_string().bold()
+    );
+    eprintln!("  {} {}", "Provider:".dimmed(), provider.name.bold());
+    eprintln!("  {} {}", "Model:".dimmed(), model.bold());
+    if api_key.is_some() {
+        eprintln!("  {} {}", "API key:".dimmed(), "••••••••".dimmed());
+    }
+    eprintln!();
+    eprintln!(
+        "{}",
+        "You're all set! Try: yaak list files in current directory".green()
+    );
+}
